Pass service config info directly to publish functions.
To do this, there needs to be an invoke function that
pthread_create() can call. Since only one argument can be passed
to the invoke function, create a wrapper structure that holds
both the service info and the weather data.  The invoke function
then calls the publish code with those arguments. This allows
us to remove the global service info array from each of the
publisher service files.