Use a function pointer table to access the service functions.
Instead of having the main code directly call into each publishing
service, have each service create a function table and attach that
to the service info record.  This removes the need for any fixed
lenght array of services.

Currently, each service needs to provide a function for
  init()    - Do any service specific initialization.
  update()  - Send the weather data to the servcie.
  cleanup() - Do any service specific clean up.

  service_setup() - Attach the above function table to the
                    service info record.

Because we're not dynamically loading each publishing service, we
still need to do a lookup between the service name and the
publishing service setup function.