Handle acurite tower sensor data.
The tower sensor data is broadcast using a JSON structure
identical to the WeatherFlow Air sensor. However, we do want
to keep them separate since they may be in very different
locations.

Tower sensors are just temperature and humidity, but the
sensor structure could be expanded to handle additional data
if needed.

Add an auxillary list of sensors to the weather data structure
and add any tower sensors to the list.  The list is dynamic so
we can support any number of addional sensors.

Also add a sensor ID to name/location mapping in the config
file so that each sensor can be more easily identified when
published.

Update the display publisher to show all the sensors.